[mesh] improved lods and lod selection, you can't spot them changing meters in front of you anymore, they are more game ready

+++ runtime/Game/Game.cpp
@@ -1592,7 +1592,7 @@ namespace spartan
                 };
             
                 generate_path(generate_path, {0, 0}, NUM_ROOMS - 1);
-                int actual_rooms = path.size();
+                int actual_rooms = static_cast<int>(path.size());
             
                 // doors
                 vector<Direction> doors(actual_rooms);
